Add efficient and general fprint comparison metrics
Expand comparison metrics for individual fingerprints to operate on
pairs of fingerprints or fingerprint databases. Array comparison
metrics used for databases are optimized to efficiently handle both
sparse and dense inputs.
Soergel similarity is optimized using Cython; however, the resulting
C-code is included so that Cython is not a required dependency.